+++ SplunkAgent/Sources/SplunkAgentObjC/Modules/Network/Model Conversions/NetworkInstrumentationConfigurationObjC+Conversions.swift
@@ -1,6 +1,6 @@
 //
 /*
-Copyright 2025 Splunk Inc.
+Copyright 2026 Splunk Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -28,7 +28,9 @@ extension NetworkInstrumentationConfigurationObjC: ModuleConfigurationSwift {
 
         return NetworkInstrumentationConfiguration(
             isEnabled: isEnabled,
-            ignoreURLs: configurationIgnoreUrls
+            ignoreURLs: configurationIgnoreUrls,
+            capturedRequestHeaders: capturedRequestHeaders,
+            capturedResponseHeaders: capturedResponseHeaders
         )
     }
 }