+++ CHANGELOG.md
@@ -7,6 +7,18 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.2.2] - 2026-04-15
+
+### Added
+
+* Added public references to sessionWillResetNotification and sessionDidResetNotification. #614
+* Added optional support to capture network headers. #604
+* Session Replay now captures text content in wireframes for UIKit applications. #615
+
+### Changed
+
+* Updated endpoint from signalfx.com to observability.splunkcloud.com. #612
+
 ## [2.2.1] - 2026-03-19
 
 ### Added